[ maintenance ]: separating graphql to new module
